start chat window after getting account
