+++ src/koma/controller/requests/account/login/login.kt
@@ -2,15 +2,14 @@ package koma.controller.requests.account.login
 
 import com.github.kittinunf.result.Result
 import com.squareup.moshi.JsonEncodingException
-import controller.LoginController
 import javafx.scene.control.Alert
+import koma.gui.view.window.auth.login.startChat
 import koma.matrix.user.identity.UserId_new
 import koma.storage.config.profile.Profile
 import koma.storage.config.server.serverConfWithAddr
 import koma.util.coroutine.adapter.retrofit.awaitMatrix
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.async
 import kotlinx.coroutines.javafx.JavaFx
 import kotlinx.coroutines.launch
 import matrix.UserPassword
@@ -20,7 +19,7 @@ import tornadofx.*
 /**
  * accept text of text fields as parameters
  */
-fun doLogin(user: String, password: String, server: String, controller: LoginController) =GlobalScope.async {
+suspend fun doLogin(user: String, password: String, server: String) {
     val userid = UserId_new(user)
     val servCon = serverConfWithAddr(userid.server, server)
     val authedProfile: Profile = if (!password.isBlank()) {
@@ -34,11 +33,11 @@ fun doLogin(user: String, password: String, server: String, controller: LoginCon
                 val message = if (ex is JsonEncodingException) {
                     "Does $server have a valid JSON API?"
                 } else mes
-                launch(Dispatchers.JavaFx) {
+                GlobalScope.launch(Dispatchers.JavaFx) {
                     alert(Alert.AlertType.ERROR, "Login Fail with Error",
                            message)
                 }
-                return@async
+                return
             }
         }
         auth
@@ -49,11 +48,9 @@ fun doLogin(user: String, password: String, server: String, controller: LoginCon
                 alert(Alert.AlertType.ERROR, "Failed to login as $userid",
                         "No access token")
             }
-            return@async
+            return
         }
         p
     }
-    launch(Dispatchers.JavaFx) {
-        controller.postLogin(authedProfile, servCon)
-    }
+    startChat(authedProfile, servCon)
 }
